ServiceWorker: size reduction via code splitting (part 1)
