+++ src/Environment.ts
@@ -1,4 +1,3 @@
-import bowser from 'bowser';
 import SdkEnvironment from './managers/SdkEnvironment';
 import { WindowEnvironmentKind } from './models/WindowEnvironmentKind';
